(uimenu) make header
