+++ include/functions.h
@@ -2487,50 +2487,6 @@ void CharStoreAI(struct Unit* unit, const struct UnitDefinition* uDef);
 // ??? sub_804DF38(???);
 // ??? sub_804E024(???);
 // ??? sub_804E03C(???);
-// ??? NewMenu_DefaultAdjusted(???);
-// ??? NewMenu_DefaultChild(???);
-// ??? sub_804EBD8(???);
-// ??? NewMenu_Default(???);
-// ??? sub_804EBF4(???);
-// ??? NewMenu_804EC34(???);
-// ??? sub_804EC58(???);
-// ??? sub_804EC74(???);
-// ??? NewMenu_BG0BG1(???);
-// ??? NewMenu(???);
-// ??? EndMenu(???);
-// ??? EndAllMenus(???);
-// ??? Menu_CallDefinedConstructors(???);
-// ??? Menu_Draw(???);
-// ??? Menu_DrawHoverThing(???);
-// ??? Menu_Idle(???);
-// ??? Menu_HandleDirectionInputs(???);
-// ??? Menu_HandleSelectInputs(???);
-// ??? Menu_GetCursorGfxCurrentPosition(???);
-// ??? MenuCommand_UsabilityAlways(???);
-// ??? UsabilityGrayed(???);
-// ??? UsabilityNever(???);
-// ??? sub_804F454(???);
-// ??? sub_804F458(???);
-// ??? sub_804F474(???);
-// ??? sub_804F4A0(???);
-// ??? sub_804F520(???);
-// ??? sub_804F530(???);
-// ??? Menu_CallTextBox(???);
-// ??? sub_804F5B4(???);
-// ??? sub_804F5FC(???);
-void MarkSomethingInMenu(void);
-void sub_804F62C(void); // bmmenu?
-// ??? NewMenu_AndDoSomethingCommands(???);
-// ??? Menu_UpdateMovingCursorGfxPosition(???);
-void ClearMenuRelatedList(void);
-// ??? sub_804F714(???);
-// ??? sub_804F754(???);
-// ??? sub_804F77C(???);
-// ??? sub_804F7AC(???);
-// ??? sub_804F7E8(???);
-// ??? sub_804F82C(???);
-// ??? sub_804F84C(???);
-// ??? sub_804F890(???);
 void InitTargets(int xRoot, int yRoot);
 void AddTarget(int x, int y, int uId, int tId);
 // ??? LinkTargets(???);
